Move some msa tests into a separate file
